+++ package.json
@@ -100,11 +100,11 @@
     "prettier": "catalog:dev",
     "publint": "catalog:dev",
     "tsup": "catalog:dev",
-    "unbuild": "catalog:prod",
     "tsx": "catalog:dev",
     "typedoc": "catalog:docs",
     "typedoc-plugin-markdown": "catalog:docs",
     "typescript": "catalog:dev",
+    "unbuild": "catalog:prod",
     "unocss": "catalog:docs",
     "unplugin-unused": "catalog:dev",
     "vite": "catalog:docs",
